fix: cicd, solve #49
