+++ docs/img/packages_coophive.dot
@@ -14,7 +14,6 @@ charset="utf-8"
 "coophive.job_offer" [color="#77AADD", label=<coophive.job_offer>, shape="box", style="filled"];
 "coophive.log_json" [color="#77AADD", label=<coophive.log_json>, shape="box", style="filled"];
 "coophive.machine" [color="#77AADD", label=<coophive.machine>, shape="box", style="filled"];
-"coophive.main_old" [color="#77AADD", label=<coophive.main_old>, shape="box", style="filled"];
 "coophive.match" [color="#77AADD", label=<coophive.match>, shape="box", style="filled"];
 "coophive.policy" [color="#77AADD", label=<coophive.policy>, shape="box", style="filled"];
 "coophive.resource_offer" [color="#77AADD", label=<coophive.resource_offer>, shape="box", style="filled"];
@@ -25,9 +24,10 @@ charset="utf-8"
 "coophive.utils" [color="#77AADD", label=<coophive.utils>, shape="box", style="filled"];
 "coophive.__main__" -> "coophive.cli" [arrowhead="open", arrowtail="none"];
 "coophive.agent" -> "coophive.deal" [arrowhead="open", arrowtail="none"];
+"coophive.agent" -> "coophive.job_offer" [arrowhead="open", arrowtail="none"];
 "coophive.agent" -> "coophive.log_json" [arrowhead="open", arrowtail="none"];
 "coophive.agent" -> "coophive.match" [arrowhead="open", arrowtail="none"];
-"coophive.agent" -> "coophive.smart_contract" [arrowhead="open", arrowtail="none"];
+"coophive.agent" -> "coophive.resource_offer" [arrowhead="open", arrowtail="none"];
 "coophive.agent" -> "coophive.utils" [arrowhead="open", arrowtail="none"];
 "coophive.cli" -> "coophive.client" [arrowhead="open", arrowtail="none"];
 "coophive.cli" -> "coophive.job_offer" [arrowhead="open", arrowtail="none"];
@@ -51,14 +51,6 @@ charset="utf-8"
 "coophive.deal" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
 "coophive.job_offer" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
 "coophive.machine" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.client" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.job" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.job_offer" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.machine" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.resource_offer" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.resource_provider" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.solver" [arrowhead="open", arrowtail="none"];
-"coophive.main_old" -> "coophive.utils" [arrowhead="open", arrowtail="none"];
 "coophive.match" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
 "coophive.resource_offer" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
 "coophive.resource_provider" -> "coophive.agent" [arrowhead="open", arrowtail="none"];
@@ -74,10 +66,8 @@ charset="utf-8"
 "coophive.smart_contract" -> "coophive.agent" [arrowhead="open", arrowtail="none"];
 "coophive.smart_contract" -> "coophive.deal" [arrowhead="open", arrowtail="none"];
 "coophive.smart_contract" -> "coophive.event" [arrowhead="open", arrowtail="none"];
-"coophive.smart_contract" -> "coophive.job_offer" [arrowhead="open", arrowtail="none"];
 "coophive.smart_contract" -> "coophive.log_json" [arrowhead="open", arrowtail="none"];
 "coophive.smart_contract" -> "coophive.match" [arrowhead="open", arrowtail="none"];
-"coophive.smart_contract" -> "coophive.resource_offer" [arrowhead="open", arrowtail="none"];
 "coophive.smart_contract" -> "coophive.result" [arrowhead="open", arrowtail="none"];
 "coophive.smart_contract" -> "coophive.utils" [arrowhead="open", arrowtail="none"];
 "coophive.solver" -> "coophive.agent" [arrowhead="open", arrowtail="none"];
